# api.yaml
openapi: 3.0.3
info:
  title: API for project 3d-inventory
  description: >-
    REST API for **3d-inventory**. Project is a simple solution that allows you to build a spatial and database
    representation of all types of warehouses and server rooms.
  contact:
    name: 3d-inventory-mongo-api
    url: https://github.com/karol-preiskorn/3d-inventory-mongo-api/
  version: 0.54.0
servers:
  - url: https://127.0.0.1:8080
  - url: http://0.0.0.0:8080
  - url: https://0.0.0.0:8080
  - url: http://localhost:8080
  - url: https://localhost:8080
  - url: http://127.0.0.1:8080
  - url: http://172.17.0.2:8080
  - url: http://172.17.0.3:8080
  - url: http://172.20.0.3:8080
  - url: https://d-inventory-api-wzwe3odv7q-ew.a.run.app
  - url: https://3d-inventory-api.ultimasolution.pl
  - url: https://3d-inventory.ultimasolution.pl
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Get all user data
      description: Get all user data
      operationId: usersGet
      security:
        - AdminAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Not found
      x-swagger-router-controller: Users
    put:
      tags:
        - users
      summary: Update user
      description: Create user
      operationId: usersPut
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Not found
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: Users
    post:
      tags:
        - users
      summary: Create new user
      description: Create new user
      operationId: userPost
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request.
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: Users
    delete:
      tags:
        - users
      summary: Delete all users
      description: Delete all users
      operationId: usersDelete
      security:
        - AdminAuth: []
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '401':
          description: Unauthorized - Missing or invalid authentication
        '403':
          description: Forbidden - Insufficient privileges (admin role required)
        '404':
          description: Not found
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: Users
  /users/{id}:
    get:
      tags:
        - users
      summary: Get the single user :id
      description: Get single user
      operationId: userGetId
      parameters:
        - name: id
          in: path
          description: ID of user
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
            pattern: ^[0-9A-Fa-f]{24}$
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found
      x-swagger-router-controller: Users
    post:
      tags:
        - users
      summary: Create a new user
      description: Create a new user
      operationId: userCreate
      parameters:
        - name: id
          required: true
          in: path
          description: ID of user
          style: simple
          explode: false
          schema:
            type: string
            example: 6566ed4ace63d42b25fbba19
            minLength: 24
            maxLength: 24
            x-42c-format: o:uuid
            pattern: ^[0-9A-Fa-f]{24}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK. Indicates that the request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmation'
        '400':
          description: Bad request.
        '500':
          description: >-
            Unexpected server error. The server encountered an unexpected condition that prevented it
            fromfulfilling the request.
      x-swagger-router-controller: Users
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication. Include role and permissions for authorization.
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with admin role required
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with user role or higher required
    ViewerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token with viewer role or higher required (read-only access)
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 6566ed4ace63d42b25fbba19
          minLength: 24
          maxLength: 24
          x-42c-format: o:uuid
          pattern: ^[0-9A-Fa-f]{24}$
        name:
          type: string
          minLength: 2
          maxLength: 256
        email:
          type: string
          format: email
          example: support@company.co.uk
          pattern: ^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
          minLength: 6
          maxLength: 320
          x-42c-format: o:email
        password:
          type: string
          minLength: 8
          maxLength: 256
        token:
          type: string
          minLength: 20
          maxLength: 256
          example: hGlyJ6KGJtKrMvNcpMwY1lRs6Cys3yCHQ0ZxMT7mCKFUOdrIOd
    Users:
      type: array
      description: Array of User
      maxItems: 1000
      items:
        $ref: '#/components/schemas/User'
...
    Confirmation:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
      description: Universal confirmation message from database
      example:
        message: Succeed
...